# deflate: Move advanced compression state