Only allocate advanced compression state when actually needed.

This will reduce upfront allocations for levels < 5 significantly (reduced by approximately 750K)

Much less risk than #70 - but not as clean.

Of course this is only relevant when `Reset()` is not used.

Benchmark: https://gist.github.com/klauspost/f5df3a3522ac4bcb3bcde448872dffe6
Before:
```
BenchmarkCompressAllocations/level(-2)/flate-8   	    5000	    306197 ns/op	 1000880 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(-2)/gzip-8    	    5000	    279000 ns/op	 1001056 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(-1)/flate-8   	    5000	    284000 ns/op	 1003184 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(-1)/gzip-8    	    5000	    312400 ns/op	 1003360 B/op	      16 allocs/op
BenchmarkCompressAllocations/level(0)/flate-8    	   10000	    177000 ns/op	  996273 B/op	      12 allocs/op
BenchmarkCompressAllocations/level(0)/gzip-8     	   10000	    207600 ns/op	  996448 B/op	      13 allocs/op
BenchmarkCompressAllocations/level(1)/flate-8    	    5000	    248399 ns/op	 1003184 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(1)/gzip-8     	   10000	    219800 ns/op	 1003360 B/op	      16 allocs/op
BenchmarkCompressAllocations/level(2)/flate-8    	    5000	    252599 ns/op	 1207984 B/op	      17 allocs/op
BenchmarkCompressAllocations/level(2)/gzip-8     	    5000	    261799 ns/op	 1208160 B/op	      18 allocs/op
BenchmarkCompressAllocations/level(3)/flate-8    	    5000	    336599 ns/op	 1339056 B/op	      17 allocs/op
BenchmarkCompressAllocations/level(3)/gzip-8     	    5000	    298599 ns/op	 1339232 B/op	      18 allocs/op
BenchmarkCompressAllocations/level(4)/flate-8    	    5000	    338400 ns/op	 1339056 B/op	      17 allocs/op
BenchmarkCompressAllocations/level(4)/gzip-8     	    5000	    298207 ns/op	 1339232 B/op	      18 allocs/op
BenchmarkCompressAllocations/level(5)/flate-8    	    5000	    273799 ns/op	 1003184 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(5)/gzip-8     	    5000	    285599 ns/op	 1003360 B/op	      16 allocs/op
BenchmarkCompressAllocations/level(6)/flate-8    	    5000	    274599 ns/op	 1003184 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(6)/gzip-8     	    5000	    300200 ns/op	 1003360 B/op	      16 allocs/op
BenchmarkCompressAllocations/level(7)/flate-8    	    5000	    277201 ns/op	 1003184 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(7)/gzip-8     	    5000	    282399 ns/op	 1003360 B/op	      16 allocs/op
BenchmarkCompressAllocations/level(8)/flate-8    	    5000	    273199 ns/op	 1003184 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(8)/gzip-8     	    5000	    278199 ns/op	 1003360 B/op	      16 allocs/op
BenchmarkCompressAllocations/level(9)/flate-8    	    5000	    273807 ns/op	 1003184 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(9)/gzip-8     	    5000	    274001 ns/op	 1003360 B/op	      16 allocs/op
```
After:
```
BenchmarkCompressAllocations/level(-2)/flate-8   	   20000	     65653 ns/op	  345520 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(-2)/gzip-8    	   20000	     66400 ns/op	  345696 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(-1)/flate-8   	    3000	    355331 ns/op	 1011377 B/op	      16 allocs/op
BenchmarkCompressAllocations/level(-1)/gzip-8    	    3000	    364333 ns/op	 1011554 B/op	      17 allocs/op
BenchmarkCompressAllocations/level(0)/flate-8    	   20000	     56000 ns/op	  340912 B/op	      12 allocs/op
BenchmarkCompressAllocations/level(0)/gzip-8     	   30000	     57966 ns/op	  341088 B/op	      13 allocs/op
BenchmarkCompressAllocations/level(1)/flate-8    	   20000	     65050 ns/op	  347824 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(1)/gzip-8     	   20000	     66250 ns/op	  348000 B/op	      16 allocs/op
BenchmarkCompressAllocations/level(2)/flate-8    	   10000	    107800 ns/op	  552628 B/op	      17 allocs/op
BenchmarkCompressAllocations/level(2)/gzip-8     	   10000	    100600 ns/op	  552804 B/op	      18 allocs/op
BenchmarkCompressAllocations/level(3)/flate-8    	   10000	    125400 ns/op	  683703 B/op	      17 allocs/op
BenchmarkCompressAllocations/level(3)/gzip-8     	   10000	    122300 ns/op	  683878 B/op	      18 allocs/op
BenchmarkCompressAllocations/level(4)/flate-8    	   10000	    120200 ns/op	  683701 B/op	      17 allocs/op
BenchmarkCompressAllocations/level(4)/gzip-8     	   10000	    116500 ns/op	  683877 B/op	      18 allocs/op
BenchmarkCompressAllocations/level(5)/flate-8    	    3000	    398334 ns/op	 1011378 B/op	      16 allocs/op
BenchmarkCompressAllocations/level(5)/gzip-8     	    5000	    384200 ns/op	 1011554 B/op	      17 allocs/op
BenchmarkCompressAllocations/level(6)/flate-8    	    5000	    440799 ns/op	 1011378 B/op	      16 allocs/op
BenchmarkCompressAllocations/level(6)/gzip-8     	    3000	    399333 ns/op	 1011553 B/op	      17 allocs/op
BenchmarkCompressAllocations/level(7)/flate-8    	    5000	    346799 ns/op	 1011377 B/op	      16 allocs/op
BenchmarkCompressAllocations/level(7)/gzip-8     	    5000	    316600 ns/op	 1011552 B/op	      17 allocs/op
BenchmarkCompressAllocations/level(8)/flate-8    	    5000	    408000 ns/op	 1011378 B/op	      16 allocs/op
BenchmarkCompressAllocations/level(8)/gzip-8     	    5000	    351599 ns/op	 1011553 B/op	      17 allocs/op
BenchmarkCompressAllocations/level(9)/flate-8    	    5000	    307600 ns/op	 1011377 B/op	      16 allocs/op
BenchmarkCompressAllocations/level(9)/gzip-8     	    5000	    317199 ns/op	 1011553 B/op	      17 allocs/op
PASS
```