@nhooyr 2020-02-15
Was this change reverted?

These are the results I'm getting on master:

```
$ go test -bench=BenchmarkCompressAllocations
goos: darwin
goarch: amd64
pkg: github.com/klauspost/compress
BenchmarkCompressAllocations/level(-2)/flate-8   	   29196	     41301 ns/op	  342272 B/op	      11 allocs/op
BenchmarkCompressAllocations/level(-2)/gzip-8    	   28676	     56400 ns/op	  342448 B/op	      12 allocs/op
BenchmarkCompressAllocations/level(-1)/flate-8   	    2926	    400765 ns/op	 3235206 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(-1)/gzip-8    	    4066	    386490 ns/op	 3235383 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(0)/flate-8    	   30462	     48217 ns/op	  339968 B/op	       9 allocs/op
BenchmarkCompressAllocations/level(0)/gzip-8     	   21348	     49214 ns/op	  340144 B/op	      10 allocs/op
BenchmarkCompressAllocations/level(1)/flate-8    	    5511	    213282 ns/op	 2186627 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(1)/gzip-8     	    5860	    212243 ns/op	 2186803 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(2)/flate-8    	    2802	    389814 ns/op	 3759528 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(2)/gzip-8     	    2666	    391217 ns/op	 3759700 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(3)/flate-8    	    4395	    291127 ns/op	 2710922 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(3)/gzip-8     	    4207	    302403 ns/op	 2711098 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(4)/flate-8    	    3687	    360733 ns/op	 2710924 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(4)/gzip-8     	    3758	    475015 ns/op	 2711098 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(5)/flate-8    	    2434	    455862 ns/op	 3235206 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(5)/gzip-8     	    2836	    366831 ns/op	 3235382 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(6)/flate-8    	    3193	    386058 ns/op	 3235205 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(6)/gzip-8     	    3214	    398171 ns/op	 3235383 B/op	      15 allocs/op
BenchmarkCompressAllocations/level(7)/flate-8    	    8619	    139871 ns/op	 1006982 B/op	      13 allocs/op
BenchmarkCompressAllocations/level(7)/gzip-8     	    8251	    151450 ns/op	 1007158 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(8)/flate-8    	   10000	    129895 ns/op	 1006982 B/op	      13 allocs/op
BenchmarkCompressAllocations/level(8)/gzip-8     	   10000	    109056 ns/op	 1007157 B/op	      14 allocs/op
BenchmarkCompressAllocations/level(9)/flate-8    	   10000	    122040 ns/op	 1006981 B/op	      13 allocs/op
BenchmarkCompressAllocations/level(9)/gzip-8     	   10000	    127054 ns/op	 1007157 B/op	      14 allocs/op
PASS
ok  	github.com/klauspost/compress	37.086s
```

Much higher than those reported in this PR.

I had similar results testing on my linux VM.