@klauspost 2020-02-15
@nhooyr This PR is about static allocations (allocations kept in the writer between writes/resets), though this seems rather high.

Could you share the benchmark code?