@nhooyr 2020-02-15
It's the same benchmark in this PR.

https://github.com/klauspost/compress/blob/5e801bfbf1025b2fbb0b3c2af9ff3c12a3677610/compressible_test.go#L243